Refactor MobileNav component - Move static data to headerData module - Refactor component
