+++ src/data/headerData.js
@@ -1,0 +1,24 @@
+const navOptions = [
+  {
+    key: 'about',
+    url: '#about',
+    icon: 'user',
+  },
+  {
+    key: 'portfolio',
+    url: '#portfolio',
+    icon: 'clipboard-list',
+  },
+  {
+    key: 'technical',
+    url: '#technical',
+    icon: 'laptop-code',
+  },
+  {
+    key: 'contact',
+    url: '#contact',
+    icon: 'envelope',
+  },
+];
+
+export default { navOptions };
